move TextPage to Pages folder
