+++ AstraBox/App.py
@@ -9,7 +9,7 @@ from AstraBox.Views.ContentFrame import ContentFrame
 from AstraBox.Views.EmptyView import EmptyView
 from AstraBox.Views.RTModelView import RTModelView
 from AstraBox.Views.TextView import TextView
-from AstraBox.Views.TextPage import TextPage
+from AstraBox.Pages.TextPage import TextPage
 from AstraBox.Views.RaceView import RaceView
 from AstraBox.Views.RunAstraView import RunAstraView
